requirements.txt
pandas
numpy
sentence-transformers
umap-learn
hdbscan
sklearn
beautifulsoup4
seaborn
matplotlib
wordcloud
pyyaml
faiss-cpu
pytz
pytest
jupyter
